Full Annotation Using Shape-constrained Trees

This function is run after running and tuning generateAnnotationThresholds in order to discover and annotate cell subsets in a cytometry experiment.

discoverPhenotypes(
  gatingSet,
  projectPath = normalizePath("."),
  nameOccuranceNum = ceiling((0.5 * length(gatingSet))),
  debugFlag = FALSE,
  threadNum = 1,
  seedValue = 123,
  densitySubSampleThreshold = 1e+06,
  densitySubSampleSize = 1e+06,
  densitySubSampleIterations = 1,
  archDescriptionList = list(targetArch = c("singleCPU")),
  plottingDevice = "pdf"
)

Arguments

gatingSet
The FAUST method requires that experimental cytometry data you wish to analyze is stored in a GatingSet data structure. This data structure can be constructed available using the flowWorkspace Bioconductor package.

projectPath
An absolute path on your system that locates all intermediate data produced by faust. At the location specified by the projectPath, faust creates a directory called faustData, and subdirectories that contain intermediate data generated by the analysis. The final output of faust, an annotated matrix of counts, is produced at projectPath/faustData/faustCountMatrix.rds, and can be loaded into R using the readRDS function.

nameOccuranceNum
The number of times a name has to appear in distinct SCAMP clusterings to be gated out.

debugFlag
Boolean value. Set to TRUE to print method status information to the console or a log file.

threadNum
Integer value. Many components of the FAUST method support multi-threading on a single CPU. Set this parameter to the number of threads you wish to use.

seedValue
Integer value that determines the random seed. Used for reproducibility.

densitySubSampleThreshold
Integer value. Sets the number of cells needed in a clustering collection to sub-sample for density estimation. NOTE: sub-sampling only occurs for density estimation. The dip test is computed on all cells in the clustering collection.

densitySubSampleSize
Integer value. The number of cells to sub-sample from a clustering collection for density estimation. NOTE: sub-sampling only occurs for density estimation. The dip test is computed on all cells in the clustering collection.

densitySubSampleIterations
Integer value. The number of distinct sub-sampled density estimates to compute. The final gate location is the median across the sub-sampled densities. NOTE: sub-sampling only occurs for density estimation. The dip test is computed on all cells in the clustering collection.

archDescriptionList
list containing slot "targetArch". Default "singleCPU" indicates FAUST will run on a single processor.

FAUST also has preliminary support for dispatching jobs across nodes of a cluster managed by slurm. This support assumes the "sbatch" command is available to launch jobs on the cluster.

Also note that the current implementation does not dynamically restart jobs that fail to launch after sbatch is invoked. In the event a job fails to launch, the faust call must be interrupted and called again.

faust attempts to check-point much of the intermediate progress in order to avoid performing redundant computation.

Set "targetArch" slot to "slurmCluster" to dispatch jobs across a cluster managed by slurm. When "targetArch" is set to "slurmCluster", the following slots must be set by the user.

Slot "partitionID": Character string. The name of the partition on which jobs are run.

Slot "jobPrefix": Character string. This string will prepend all jobs launched on the slurm cluster. When "squeue" is used to interrogate job state, this string can be used to identify jobs.

Slot "jobTime"" Character string. This string sets the maximum time a job launched by faust on the slurm cluster can run. String format follows 'HH:MM:SS'.

Slot "maxNodeNum": Numeric value. The maximum number of nodes you wish to request on the slurm cluster.

Slot "maxTime": Numeric value. The total amount of time you want to run FAUST for. When exceeded, the job terminates.

Slot "nodeThreadNum": Character string. The number of threads used by FAUST on each node.

Slot "sbatchFlags": Character string. This string contains space-delimited command-line flags to pass to sbatch.

plottingDevice
string with device for saving graphical output. By default it is set to "pdf".

Value

discoverPhenotypes returns a null value on completion. The main output is the file "projectPath/faustData/faustCountMatrix.rds". The rownames are sampleNames(gatingSet)] and the column names are the cell populations discovered by faust. Note that the special cell population "0_0_0_0_0" counts unclassified cells in the experiment.

Examples

#Please see the vignette "discoverPhenotypes" for an introduction on using FAUST.